% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ase.R
\name{ase}
\alias{ase}
\title{Adjacency Spectral Embedding}
\usage{
ase(A, p = 2L, q = 0L, scale = TRUE)
}
\arguments{
\item{A}{The adjacency matrix of the graph to be embedded}

\item{p}{Number of assortative dimensions}

\item{q}{Number of disassortative dimensions}

\item{scale}{Whether to scale the embedding by the eigenvalues}
}
\value{
An nrow(A) by (p+q) matrix of embedded points
}
\description{
Adjacency spectral embedding of undirected graph with adjacency matrix A.
}
\examples{
# generate a small homogeneous SBM with two communities
n1 <- 20
n2 <- 20
n <- n1 + n2
p.within <- .5
p.between <- .1
P <- matrix(p.between, nrow = n, ncol = n)
P[seq(n1), seq(n1)] <- p.within
P[seq(n1 + 1, n), seq(n1 + 1, n)] <- p.within
A <- osc::draw.graph(P)

# embed A in two assortative dimensions
X <- ase(A, 2, 0)
# the embedding will consist of points around two point masses
plot(X)
}
