% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity_matrix.R
\name{compute.affinity.matrix}
\alias{compute.affinity.matrix}
\title{Compute the affinity matrix for orthogonal spectral clustering}
\usage{
compute.affinity.matrix(A, K = 2L, model = "pabm", assortative = TRUE)
}
\arguments{
\item{A}{Adjacency matrix describing an undirected graph.}

\item{K}{Number of communities.}

\item{model}{Type of block model (SBM, DCBM, or PABM).
The dimensionality of the embedding of A depends on this model.
If the model is SBM or DCBM, then the embedding dimension is K.
If the model is PABM, then the embedding dimension is \eqn{K^2}
and \eqn{p = K (K + 1) / 2} and \eqn{q = K (K - 1) / 2}.}

\item{assortative}{Whether the block model is assortative or disassortative.
Only valid for SBM and DCBM. This argument is ignored for PABM.
If assortative is TRUE, then p = K and q = 0.
If assortative is FALSE, then p = 1 and q = K - 1.
If assortative is an integer, then p is the value of that integer and
q = K - p.}
}
\value{
B, the affinity matrix.
\eqn{B[i, j]} approaches 0 asymptotically
if i and j are in different communities.
V, first p and last q eigenvectors of A.
The rows of V that correspond to vertices in two different communities
approach orthogonality asymptotically.
p, q, the embedding dimensions.
}
\description{
Compute the affinity matrix for orthogonal spectral clustering
}
\examples{
# generate a small homogeneous SBM with two communities
n1 <- 20
n2 <- 20
n <- n1 + n2
p.within <- .5
p.between <- .1
P <- matrix(p.between, nrow = n, ncol = n)
P[seq(n1), seq(n1)] <- p.within
P[seq(n1 + 1, n), seq(n1 + 1, n)] <- p.within
A <- orthoclust::draw.graph(P)

# compute the affinity matrix
B <- orthoclust::compute.affinity.matrix(A, K = 2, model = 'sbm')
}
