% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pabm_p.R
\name{pabm.edge.probability.matrix}
\alias{pabm.edge.probability.matrix}
\title{Construct the edge probability matrix for a PABM}
\usage{
pabm.edge.probability.matrix(groups, popularity.params)
}
\arguments{
\item{groups}{A vector of group labels.}

\item{popularity.params}{An n by K matrix of popularity parameters.
\expr{popularity.params\link{i, k}} is vertex i's affinity to community k.}
}
\value{
The n by n edge probability matrix.
}
\description{
Construct the edge probability matrix for a PABM
}
\examples{
# set the size of each group
n1 <- 100
n2 <- 100
n3 <- 100
n <- n1 + n2 + n3
}
